CheckDlgButton, CheckRadioButton, IsDlgButtonChecked: safe import
The calls to CheckDlgButton, CheckRadioButton, and IsDlgButtonChecked must be 'safe' as they send messages to the controls, which will be handled by the calling application's window procedure (if it has one) while in the call to the function.  This may trigger Haskell code to be called before the foreign function returns.