EnableWindow: safe import
The call to EnableWindow must be 'safe' as it sends a WM_CANCELMODE and/or a WM_ENABLE message before it returns.  These will be handled by the calling application's window procedure (if it has one) while in the call to EnableWindow.  This may trigger Haskell code to be called before the foreign function returns.