IsDlgButtonChecked: return value of 0 is not an error
A return value of 0 from IsDlgButtonChecked is a valid button state (BST_UNCHECKED).  In addition, this function does not define a return value for failure, and it does not specify to use GetLastError to determine if an error occured.  

See http://msdn.microsoft.com/en-us/library/windows/desktop/bb761879%28v=vs.85%29.aspx